+++ src/data/print_documents/options.py
@@ -24,6 +24,7 @@ from data.print_documents.player_sorters import (
 )
 from data.print_documents.player_splitters import PlayerSplitter, NoSplitPlayerSplitter
 from data.print_documents.qrcode_types import NetworkQRCodeType, QRCodeType
+from data.print_documents.team_types import TeamType, ClubTeamType
 from utils.option import Option
 
 if TYPE_CHECKING:
@@ -670,6 +671,128 @@ class PlayerHistoryOption(PrintOption):
     @staticmethod
     def static_id() -> str:
         return 'player-history'
+
+    @property
+    def type(self) -> type | UnionType:
+        return bool
+
+    @property
+    def default_value(self) -> Any:
+        return False
+
+
+class TeamTypePrintOption(PrintOption):
+    @staticmethod
+    def static_id() -> str:
+        return 'team-type'
+
+    @property
+    def type(self) -> type | UnionType:
+        return str
+
+    @property
+    def default_value(self) -> Any:
+        return ClubTeamType.static_id()
+
+    @property
+    def team_type_options(self) -> dict[str, str]:
+        from data.print_documents import PrintTeamTypeManager
+
+        return PrintTeamTypeManager(self.event).options()
+
+    @cached_property
+    def team_type(self) -> TeamType:
+        from data.print_documents import PrintTeamTypeManager
+
+        return PrintTeamTypeManager(self.event).get_object(self.value)
+
+    @override
+    def validate(self):
+        try:
+            _type = self.team_type
+        except KeyError:
+            # Untranslated, should not happen
+            raise OptionError(f'Unknown team type: {self.value}', self)
+
+
+class TeamSizePrintOption(PrintOption):
+    @staticmethod
+    def static_id() -> str:
+        return 'team-size'
+
+    @property
+    def type(self) -> type | UnionType:
+        return int | None
+
+    @property
+    def default_value(self) -> Any:
+        return 4
+
+    @override
+    def validate(self):
+        super().validate()
+        if self.value is not None and self.value < 2:
+            raise OptionError(_('An integer greater than 1 is expected.'), self)
+
+
+class MaxTeamsPerEntityPrintOption(PrintOption, ABC):
+    @staticmethod
+    def static_id() -> str:
+        return 'max-teams-per-entity'
+
+    @property
+    def type(self) -> type | UnionType:
+        return int | None
+
+    @property
+    def default_value(self) -> Any:
+        return None
+
+    @override
+    def validate(self):
+        super().validate()
+        if self.value is not None and self.value < 1:
+            raise OptionError(_('A positive integer is expected.'), self)
+
+
+class MinGenderCountPrintOption(PrintOption):
+    @staticmethod
+    def static_id() -> str:
+        return 'min-gender-count'
+
+    @property
+    def type(self) -> type | UnionType:
+        return int | None
+
+    @property
+    def default_value(self) -> Any:
+        return None
+
+    @override
+    def validate(self):
+        super().validate()
+        if self.value is not None and self.value < 0:
+            raise OptionError(_('A positive integer is expected.'), self)
+
+
+class DisplayIncompleteTeamsPrintOption(PrintOption):
+    @staticmethod
+    def static_id() -> str:
+        return 'display-incomplete-teams'
+
+    @property
+    def type(self) -> type | UnionType:
+        return bool
+
+    @property
+    def default_value(self) -> Any:
+        return True
+
+
+class RankIncompleteTeamsFirstPrintOption(PrintOption):
+    @staticmethod
+    def static_id() -> str:
+        return 'rank-incomplete-teams-first'
 
     @property
     def type(self) -> type | UnionType: